requirements.txt
asgiref==3.3.1
Django==2.2.19
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
Markdown==3.3.4
PyJWT==2.0.1
pytz==2021.1
sqlparse==0.4.1
